WALS added. Lincoln running well. Working on growth logic.

+++ scripts/run_BASGRA_Scott.r
@@ -14,7 +14,7 @@
   # run model for this period 
   # Scott weather data is 2010-1 to 2017-318 (14 Nov)
   year_start     <- as.integer(2011)
-  doy_start      <- as.integer(152) # 1 June
+  doy_start      <- as.integer(244) # 1 Sept
   year_stop      <- as.integer(2014)
   doy_stop       <- as.integer(151) # 31 May
